requirements.txt
click>=5.0,<8.0
jsonschema==2.6.0
mock==2.0.0
pip>=9.0.1
pyflakes==1.6.0
pytest==3.2.3
pytest-cov==2.5.1
pytest-flakes==2.0.0
pyyaml==3.12
tox==2.9.1
twine==1.9.1
twisted==17.9.0
wheel==0.30.0
